requirements.txt
flask
flask-cors
cryptography
pandas
openpyxl
python-docx
PyPDF2
gunicorn
python-docx
PyPDF2